+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :feature:`11019` Users can now use exclusion filters for event type, event subtype, counterparty, location, and validator index in the history events table.
 * :feature:`1606` Users can now query their Kraken Futures balances by providing Futures API keys in the exchange settings.
 * :bug:`11240` Failed backend startups will now surface a clear error in the UI without showing raw errors.
 * :feature:`11231` Users can now click a location in the blockchain balance summary on the dashboard to filter by that chain.